Expose 128-bit color/depth read/write paths in Device abstraction
Again more optimizations for sim since we move more data per cycle, and thus don't have to simulate as many cycles during these periods.

+++ misc/strugl/src/device.rs
@@ -1,9 +1,9 @@
 pub trait Device {
     fn write_reg(&mut self, addr: u32, data: u32);
     fn read_reg(&mut self, addr: u32) -> u32;
-    fn write_color_buffer_word(&mut self, addr: u32, data: u32);
-    fn read_color_buffer_word(&mut self, addr: u32) -> u32;
-    fn write_depth_buffer_word(&mut self, addr: u32, data: u16);
-    fn read_depth_buffer_word(&mut self, addr: u32) -> u16;
+    fn write_color_buffer_word(&mut self, addr: u32, data: u128);
+    fn read_color_buffer_word(&mut self, addr: u32) -> u128;
+    fn write_depth_buffer_word(&mut self, addr: u32, data: u128);
+    fn read_depth_buffer_word(&mut self, addr: u32) -> u128;
     fn write_tex_buffer_word(&mut self, addr: u32, data: u128);
 }
